Add files via upload
This is Task 3 of my internship at TXON in Web Development and Design. In this task, I have created a to-do list App  of a list of what we do using HTML , CSS & js. A to-do list App is one of the best web development projects for beginners. This project demands a foundational understanding of HTML , CSS & js. Through this task, I understood the basics of HTML , CSS & js  i.e how to add columns, add footers, arrange text and images. Most importantly, I was able to use my creativity to make the page look impressive. The alignments, the padding, the color palette, boxes and all the other elements on the page require attention.